+++ examples/bond/test/bond-token-flow.js
@@ -16,28 +16,24 @@ const { buyTxRuntime, issueTx } = require('../scripts/run/common/common');
  * Test for the scenario described in Readme.md
  */
 describe('Bond token Tests', function () {
-  const master = new AccountStore(1000e6);
-  let appManager = new AccountStore(initialBalance);
-  let bondTokenCreator = new AccountStore(initialBalance);
   let issuerAddress = new AccountStore(minBalance);
-  let elon = new AccountStore(initialBalance);
-  let bob = new AccountStore(initialBalance);
-  let dex1 = new AccountStore(initialBalance);
-  let dex2 = new AccountStore(initialBalance);
+  const master = new AccountStore(1000e6);
+  let appManager;
+  let bondTokenCreator;
+  let elon;
+  let bob;
+  let dex1;
+  let dex2;
 
-  let runtime;
+  const runtime = new Runtime([master, issuerAddress]);
+  [appManager, bondTokenCreator, elon, bob, dex1, dex2] = runtime.defaultAccounts();
+
   let flags;
   let applicationId;
   let issuerLsigAddress;
   let lsig;
 
-  this.beforeAll(async function () {
-    runtime = new Runtime([
-      appManager, bondTokenCreator,
-      issuerAddress, master, elon,
-      bob, dex1, dex2
-    ]);
-
+  this.beforeAll(() => {
     flags = {
       sender: appManager.account,
       localInts: 1,
@@ -51,13 +47,8 @@ describe('Bond token Tests', function () {
 
   // fetch latest account state
   function syncAccounts () {
-    appManager = runtime.getAccount(appManager.address);
-    bondTokenCreator = runtime.getAccount(bondTokenCreator.address);
     issuerAddress = runtime.getAccount(issuerAddress.address);
-    elon = runtime.getAccount(elon.address);
-    dex1 = runtime.getAccount(dex1.address);
-    dex2 = runtime.getAccount(dex2.address);
-    bob = runtime.getAccount(bob.address);
+    [appManager, bondTokenCreator, elon, bob, dex1, dex2] = runtime.defaultAccounts();
   }
 
   // Bond-Dapp initialization parameters
@@ -150,12 +141,12 @@ describe('Bond token Tests', function () {
     // opt-in to app
     runtime.optInToApp(appManager.address, applicationId, {}, {});
     runtime.optInToApp(issuerAddress.address, applicationId, {}, {});
+
     syncAccounts();
     assert.isDefined(appManager.appsLocalState.get(applicationId));
     assert.isDefined(issuerAddress.appsLocalState.get(applicationId));
 
     optInLsigToBond(runtime, lsig, currentBondIndex, appManager);
-
     // Issue tokens to issuer from bond token creator
     let groupTx = issueTx(bondTokenCreator.account, lsig, applicationId, currentBondIndex);
     runtime.executeTx(groupTx);